PROGAMING-main-3/Django/Progaming/requirements.txt
django
pillow
django-admin-interface
django-rest-framework